Add support for gstreamer 1.16 by putting 1.18 specific stuff behind a build tag. For backward compatibility, when building without any build tags, the latest gstreamer version is assumed.
